+++ src/main/java/lotto/constants/Ranking.java
@@ -1,4 +1,4 @@
-package lotto.constant;
+package lotto.constants;
 
 public enum Ranking {
