+++ src/main/java/lotto/dto/CountDto.java
@@ -1,4 +1,4 @@
-package lotto.domain.dto;
+package lotto.dto;
 
 import java.util.List;
