+++ src/main/java/lotto/dto/LottosDto.java
@@ -1,7 +1,7 @@
-package lotto.domain.dto;
+package lotto.dto;
 
-import lotto.domain.model.Lotto;
-import lotto.domain.model.Lottos;
+import lotto.model.Lotto;
+import lotto.model.Lottos;
 
 import java.util.List;
