+++ src/main/java/lotto/dto/ResultDto.java
@@ -1,4 +1,4 @@
-package lotto.domain.dto;
+package lotto.dto;
 
 public record ResultDto(Integer[] winningLottoCount, Double rateOfReturn) {
 }
