+++ src/main/java/lotto/factory/LottoFactory.java
@@ -1,13 +1,13 @@
 package lotto.factory;
 
 import camp.nextstep.edu.missionutils.Randoms;
-import lotto.domain.model.Lotto;
+import lotto.model.Lotto;
 
 import java.util.List;
 
-import static lotto.constant.Number.MAX_COUNT;
-import static lotto.constant.Number.RANGE_END;
-import static lotto.constant.Number.RANGE_START;
+import static lotto.constants.Number.MAX_COUNT;
+import static lotto.constants.Number.RANGE_END;
+import static lotto.constants.Number.RANGE_START;
 
 public class LottoFactory {
     public Lotto issueLotto() {
