+++ src/test/java/lotto/model/LottosTest.java
@@ -1,4 +1,4 @@
-package lotto.domain.model;
+package lotto.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
